Добавлена возможность открывать объявление

+++ src/main/java/ru/skypro/homework/dto/AdsDto.java
@@ -24,7 +24,7 @@ public class AdsDto {
     String image;
 
     // id объявления
-    int pk;
+    long pk;
     int price;
     String title;
